Rename train.py to train_entry_point.py to avoid import names conflict
