+++ src/tf_container/__init__.py
@@ -11,7 +11,7 @@
 #  express or implied. See the License for the specific language governing 
 #  permissions and limitations under the License.
 
-from tf_container.train import train
+from tf_container.train_entry_point import train
 from tf_container.serve import load_dependencies, transformer
 import tf_container.serve as serve
